requirements.txt
# Fashion Flux Backend Dependencies
# Install with: pip install -r requirements.txt

# FastAPI framework
fastapi>=0.104.0

# ASGI server
uvicorn[standard]>=0.24.0

# Machine Learning - Use CPU version for faster deployment
tensorflow-cpu>=2.10.0,<2.16.0
numpy>=1.23.0,<2.0.0
scikit-learn>=1.2.0

# Image processing
Pillow>=9.0.0

# Data validation
pydantic>=2.0.0

# File handling
python-multipart>=0.0.6

# Production server
gunicorn>=21.0.0

